feat: rename deta to autocapture
