+++ packages/plugin-autocapture-browser/jest.config.js
