+++ packages/plugin-autocapture-browser/rollup.config.js
@@ -1,6 +1,6 @@
 import { iife, umd } from '../../scripts/build/rollup.config';
 
 iife.input = umd.input;
-iife.output.name = 'amplitudeDefaultEventTrackingAdvancedPlugin';
+iife.output.name = 'amplitudeAutocapturePlugin';
 
 export default [umd, iife];
