+++ packages/plugin-autocapture-browser/src/helpers.ts
