+++ packages/plugin-autocapture-browser/src/libs/finder.ts
