+++ packages/plugin-autocapture-browser/src/typings/autocapture.ts
