+++ packages/plugin-autocapture-browser/src/version.ts
