+++ packages/plugin-autocapture-browser/test/setup.ts
