+++ packages/plugin-autocapture-browser/test/utils.ts
