+++ packages/plugin-autocapture-browser/tsconfig.es5.json
