+++ packages/plugin-autocapture-browser/tsconfig.esm.json
