+++ packages/plugin-autocapture-browser/tsconfig.json
